criando_chat_bot/requirements.txt
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.3.0
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
